project/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
groq==0.4.1
python-dotenv==1.0.0
pydantic==2.4.2

# RAG and ML dependencies  
sentence-transformers==2.2.2
chromadb==0.4.22
langchain-community==0.0.38

# Data processing
pandas==2.1.4
numpy==1.24.4
tqdm==4.66.1

# ML frameworks
torch==2.1.2
transformers==4.36.2

# Web frontend (Flask)
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1
requests>=2.28.0
